refactor: split FileSystemState into FilesystemQuery + HistoryQuery (UPI 052 PR 1)
Narrow the 14-method plan::FileSystemState read-side trait into two traits
split along the ADR-102 axis (filesystem is truth, SQLite is history):
FilesystemQuery (snapshot listings, mounts, pin files, free space, btrfs
generation) and HistoryQuery (send sizes, calibration, send/drive timestamps),
both in a new observation.rs module.

A FileSystemState bridge supertrait + blanket impl preserves all 27 callers
and ~200 mock setups unchanged — this PR is purely additive: zero behavior
change, zero dead code, zero #[allow]. The RealFileSystemState and
MockFileSystemState impls are split into the two halves verbatim. calibrate.rs
and verify.rs gain a FilesystemQuery import because concrete-receiver method
resolution needs the declaring trait in scope; trait-object callers dispatch
through the bridge unchanged.

No on-disk (ADR-105) or config-schema (ADR-111) change; no cross-repo impact.

Modules: plan, observation, calibrate, verify
